requirements.txt
django>=3.2.12
djangorestframework>=3.9.0
uwsgi>=2.0.18
